Security improvements & more detailed error output
We prevent LaTeX from reading/writing files in parent directories;
additionally, we have introduced a list of unsupported commands which
the user will be informed on if they attempt to use any, so that they
receive a friendly error message instead of LaTeX likely encountering
an error. Finally, we have resturctured the command definitions to
use methods and multi-line blocks, for readability.